Ajustes na organização de diretórios e foco principal em assets e na Home Screen, adotando componentes e facilitando a leitura do código.
