+++ package.json
@@ -23,6 +23,7 @@
   "devDependencies": {
     "@babel/core": "^7.25.2",
     "@types/react": "~18.3.12",
+    "babel-plugin-module-resolver": "^5.0.2",
     "react-native-svg-transformer": "^1.5.0",
     "typescript": "^5.3.3"
   },
